python/packages/botas/pyproject.toml
[project]
name = "botas"
version = "0.1.0"
dependencies = [
    "httpx",
    "pydantic",
    "PyJWT",
    "cryptography",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
